Use the Windows FontFamily implementation on both Windows and Unix
