+++ src/System.Drawing.Common/src/System.Drawing.Common.csproj
@@ -139,6 +139,7 @@
     <Compile Include="System\Drawing\Internal\GPRECT.cs" />
     <Compile Include="System\Drawing\Internal\GPRECTF.cs" />
     <Compile Include="System\Drawing\Brush.cs" />
+    <Compile Include="System\Drawing\FontFamily.cs" />
     <Compile Include="System\Drawing\SolidBrush.cs" />
     <Compile Include="System\Drawing\SystemBrushes.cs" />
     <Compile Include="System\Drawing\TextureBrush.cs" />
@@ -206,7 +207,6 @@
     <Compile Include="System\Drawing\Drawing2D\GraphicsPath.cs" />
     <Compile Include="System\Drawing\Font.cs" />
     <Compile Include="System\Drawing\Font.NotSerializable.cs" />
-    <Compile Include="System\Drawing\FontFamily.cs" />
     <Compile Include="System\Drawing\GdiplusNative.Windows.cs" />
     <Compile Include="System\Drawing\Graphics.Windows.cs" />
     <Compile Include="System\Drawing\GraphicsContext.cs" />
@@ -299,7 +299,6 @@
     <Compile Include="Unix\System.Drawing\ComIStreamWrapper.cs" />
     <Compile Include="Unix\System.Drawing\ContentAlignment.cs" />
     <Compile Include="Unix\System.Drawing\Font.cs" />
-    <Compile Include="Unix\System.Drawing\FontFamily.cs" />
     <Compile Include="Unix\System.Drawing\GraphicsUnit.cs" />
     <Compile Include="Unix\System.Drawing\gdipEnums.cs" />
     <Compile Include="Unix\System.Drawing\gdipFunctions.cs" />
